requirements.txt
Django==2.2.1
djangorestframework==3.9.4
Pillow==6.0.0
pkg-resources==0.0.0
pytz==2019.1
serpy==0.3.1
six==1.12.0
sqlparse==0.3.0
